simplify createStore stuff in veritone-widgets
